Fixed effects with window splice

Takes a time series of prices and id's and returns an index using the FEWS (Fixed Effects Window Splice) method or also known as a TPD (Time Product Dummy) index.

FEWS(times, logprice, id, window_length, weight = NULL,
  splice_pos = "mean", num_cores = NULL)

Arguments

times
vector of the times at which price observations were made.
NOTE: times must be of class Date or numeric.

logprice
vector of log of prices at the given time

id
vector of distinct identification number of consumer goods

window_length
single number for length of windows of the data that regressions are fit on

weight
vector of expenditure weights used in the regressions

splice_pos
The positon on which to splice the windows together. This can be a number from 1 to window_length or any of c("window", "half","movement", "mean").

num_cores
Number of cores to use for parrallell computaion. Convention is parallel::detectCores()-1 on local machines

Value

The function retuns a list of 2 items:

fews

a dataframe of the FEWS index

fixed_effects

a dataframe of the unspliced coefficients of the fixed effects model. The number of rows in the data frame is window_length * the number of windows in the data

diagnostics

a dataframe of diagnostic results produced during model calcaulation. The diagnostics dataframe has the following columns:

Details

The function takes vectors for each of the inputs. It is important to note that the times argument muct be of numeric or Date class. This is because the order of the dates matters.

The function also has the cabpabililty to run in parallell, using the num_cores argument. Note that for smaller datasets using non-parallell code is often faster than using parralellisation due to the overhead associated with dividing the job across multiple processors.

Examples

FEWS(times = turvey$month,
logprice = log(turvey$price),
id = turvey$commodity,
window_length = 5,
weight = turvey$price * turvey$quantity,
splice_pos = "mean",
num_cores = NULL)

#> Error in FEWS(times = turvey$month, logprice = log(turvey$price), id = turvey$commodity,     window_length = 5, weight = turvey$price * turvey$quantity,     splice_pos = "mean", num_cores = NULL): could not find function "FEWS"

FEWS(times = turvey$month,
logprice = log(turvey$price),
id = turvey$commodity,
window_length = 5,
weight = turvey$price * turvey$quantity,
splice_pos = "mean",
num_cores = 2)

#> Error in FEWS(times = turvey$month, logprice = log(turvey$price), id = turvey$commodity,     window_length = 5, weight = turvey$price * turvey$quantity,     splice_pos = "mean", num_cores = 2): could not find function "FEWS"